requirements.txt
# Define your dependencies here

# Install the latest version of Django framework 4.0.2 but less than 4.1.0 version.
Django>=4.0.2,<4.1.0

# Install the latest version of Django REST framework 3.13 but less than 3.14 version.
djangorestframework>=3.13.0,<3.14.0

# Install python package 'psycopg2' that used by Django as database driver to communicate with Postgres
# server installed on Docker by using another dependency 'postgresql-client' for the Docker image.
# Note: going to have some dependencies 'build-base postgresql-dev musl-dev' to be installed by 'Dockerfile'
#       when building the image in order to install the 'postgresql-client'.
psycopg2>=2.9.3,<2.10.0

# Install Python Imaging Library that adds image create and processing capabilities to your Python interpreter.
Pillow>=9.0.1,<9.1.0

# Install python package contains a collection of tools for internationalizing Python applications, that required for
# 'PhoneNumberField' widgets.
Babel>=2.9.1,<2.10.0

# Install Django library which interfaces with 'python-phonenumbers' module to validate, pretty print and convert phone
# numbers.
django-phonenumber-field>=6.1.0,<6.2.0
django-phonenumbers>=1.0.1,<1.1.0

# Install python linting tool 'flake8'.
flake8>=4.0.1,<4.1.0